extract filter logic to hook
